Modified the OmniBeat VS solution. 	- Added a dummy Person class and played around binding the data in the person object with the Content of "Button3". Display the name.
Also, I added another directory "Non_IS_projects" and added a project that is not a InteractiveSpaceApp. Feel free to use this VS solution
and modify it in any way for learning purposes since you cannot run the "OmniBeat" project UNLESS the Kinect and Projector is connected.

Feel free to add additional VS Solution projects for the sake of learning WPF.